Deprivation deciles

Outputs a data frame allocating deprivation decile to area code based on the Indices of Multiple Deprivation (IMD) produced by Department of Communities and Local Government

Usage

deprivation_decile(
  AreaTypeID,
  Year = 2019,
  proxy_settings = fingertips_proxy_settings(),
  path
)

Arguments

AreaTypeID

Integer value; this function uses the IndicatorIDs 91872, 93275 and 93553, please use the indicator_areatypes() function to see what AreaTypeIDs are available

Year

Integer value, representing the year of IMD release to be applied, limited to 2015 or 2019

proxy_settings

string; whether to use Internet Explorer proxy settings "default" or "none". Setting this manually will decrease runtime; default determined automatically.

path

String; Fingertips API address. Function will default to the correct address

Value

A lookup table providing deprivation decile and area code

Details

This function uses the fingertips_data function to filter for the Index of multiple deprivation score for the year and area supplied, and returns the area code, along with the score and the deprivation decile, which is calculated using the ntile function from dplyr

See also

Examples

if (FALSE) {
# Return 2019 deprivation scores for Sustainability and Transformation Footprints
deprivation_decile(120, 2019)}