Indicator metadata

Outputs a data frame containing the metadata for selected indicators. Note, this function can take up to a few minutes to run (depending on internet connection speeds)

indicator_metadata(IndicatorID = NULL, DomainID = NULL, ProfileID = NULL, path)

Arguments

IndicatorID
Numeric vector, id of the indicator of interest. Also accepts "All".

DomainID
Numeric vector, id of domains of interest

ProfileID
Numeric vector, id of profiles of interest. Indicator polarity can vary between profiles therefore if using one of the comparison fields it is recommended to complete this field as well as IndicatorID. If IndicatorID is populated, ProfileID can be ignored or must be the same length as IndicatorID (but can contain NAs).

path
String; Fingertips API address. Function will default to the correct address

Value

The metadata associated with each indicator/domain/profile identified

See also

Examples

if (FALSE) {
# Returns metadata for indicator ID 90362 and 1107
indicatorIDs <- c(90362, 1107)
indicator_metadata(indicatorIDs)

# Returns metadata for the indicators within the domain 1000101
indicator_metadata(DomainID = 1000101)

# Returns metadata for the indicators within the profile with the ID 129
indicator_metadata(ProfileID = 129)}