Indicator order number

Outputs a tibble of indicator ids and their sequence number for the provided domain and area type. This enables the user to order the indicators as they are ordered on the Fingertips website.

indicator_order(DomainID, AreaTypeID, ParentAreaTypeID, path)

Arguments

DomainID
Numeric vector, id of domains of interest

AreaTypeID
Numeric vector, the Fingertips ID for the area type. This argument accepts "All", which returns data for all available area types for the indicator(s), though this can take a long time to run

ParentAreaTypeID
Numeric vector, the comparator area type for the data extracted; if NULL the function will use the first record for the specified `AreaTypeID` from the area_types() function

path
String; Fingertips API address. Function will default to the correct address

Value

A data frame of indicator ids and sequence number

See also

Examples

if (FALSE) {
indicator_order(DomainID = 1938133161, AreaTypeID = 102, ParentAreaTypeID = 6)}