Nearest neighbours

Outputs a character vector of similar areas for a given area.

Details

For new upper tier local authorities (post 4/23) the NHSE Nearest Neighbours Model is used.

All other local authorities types (lower tier and older upper tier) use the CIPFA's Nearest Neighbours Model. Older local authority geography types will use older versions of the model.

Similar areas for Clinical Commissioning Groups are based on NHS England's similar CCG explorer tool.

Currently the function does not have the ability to use the method from the Children's services statistical neighbour benchmarking tool.

Usage

nearest_neighbours(
  AreaCode,
  AreaTypeID,
  proxy_settings = fingertips_proxy_settings(),
  path
)

Arguments

AreaCode

Character vector, ONS area code of area of interest

AreaTypeID

AreaTypeID of the nearest neighbours (see nearest_neighbour_areatypeids) for available IDs

proxy_settings

string; whether to use Internet Explorer proxy settings "default" or "none". Setting this manually will decrease runtime; default determined automatically.

path

String; Fingertips API address. Function will default to the correct address

Value

A character vector of area codes

See also

Examples

if (FALSE) { # \dontrun{
nearest_neighbours(AreaCode = "E09000004", AreaTypeID = 502)} # }