Nearest neighbours

Outputs a character vector of similar areas for given area. Currently returns similar areas for Clinical Commissioning Groups (old and new) based on NHS England's similar CCG explorer tool or lower and upper tier local authorities based on CIPFA's Nearest Neighbours Model or upper tier local authorities based on Children's services statistical neighbour benchmarking tool

Usage

nearest_neighbours(
  AreaCode,
  AreaTypeID,
  measure,
  proxy_settings = fingertips_proxy_settings(),
  path
)

Arguments

AreaCode

Character vector, ONS area code of area of interest

AreaTypeID

AreaTypeID of the nearest neighbours (see nearest_neighbour_areatypeids) for available IDs

measure

deprecated. Previously a string; when AreaTypeID = 102 measure must be either "CIPFA" for CIPFA local authority nearest neighbours or "CSSN" for Children's services statistical neighbours

proxy_settings

string; whether to use Internet Explorer proxy settings "default" or "none". Setting this manually will decrease runtime; default determined automatically.

path

String; Fingertips API address. Function will default to the correct address

Value

A character vector of area codes

See also

Examples

if (FALSE) {
nearest_neighbours(AreaCode = "E38000002", AreaTypeID = 154)}